Remove unused files and refactor error handling: consolidate severity constants and clean up validation logic
