+++ .github/pull_request_template.md
@@ -1,8 +1,5 @@
 ## In This PR
+
 <!-- Please provide a brief description of the change being made -->
-- 
 
-### Related Issue(s)
-<!-- If this pull request addresses one or more issues, link them here. Example:
-Closes #123 -->
-
+- [ ]
