+++ scripts/e2e.sh
@@ -19,15 +19,3 @@ if [ $? -ne 0 ]; then
   echo "Server tests failed."
   exit 1
 fi
-
-echo "Running Python Tests..."
-# Navigate to the Python root directory before running the Python tests
-cd "$(pwd)/server/bundled/tools/python"
-python -m unittest discover -s tests
-
-if [ $? -ne 0 ]; then
-  echo "Python tests failed."
-  exit 1
-fi
-
-echo "All tests passed."
