+++ server/bundled/tools/python/util.py
@@ -10,16 +10,3 @@ def serialize_file_data(obj):
         return {k: serialize_file_data(v) for k, v in obj.items()}
     else:
         return str(obj)
-    
-
-def evaluate_expr_as_string(value: ast.expr) -> str:
-    """
-    Helper function to evaluate an AST expression as a string.
-    Safely handles exceptions and returns the lowercased string value.
-    If evaluation fails, returns an empty string.
-    """
-    try:
-        value_str = ast.literal_eval(value).strip().lower()
-        return value_str
-    except (ValueError, SyntaxError):
-        return ''
