--- server/src/commentConventions.ts
@@ -1,3 +1,0 @@
-export interface CommentConventions {
-	flagRedundant: boolean;
-}
