+++ server/src/constants/errors.ts
@@ -1,3 +1,17 @@
 export enum ERROR_CODES  {
     UNAUTHENTICATED = 403,
 }
+
+export class RateLimitError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'RateLimitError';
+    }
+}
+
+export class ForbiddenError extends Error {
+    constructor(message: string) {
+        super(message);
+        this.name = 'ForbiddenError';
+    }
+}
