--- server/src/constants/filepaths.ts
@@ -1,1 +1,0 @@
-export const PYTHON_DIRECTORY = "bundled/tools/python";
