+++ server/src/constants/severity.ts
@@ -1,0 +1,7 @@
+
+export enum Severity {
+    HINT = 'HINT',
+    INFORMATION = 'INFORMATION',
+    WARNING = 'WARNING',
+    ERROR = 'ERROR',
+}
