--- server/src/data/verbs.ts
@@ -1,234 +1,0 @@
-export const verbDictionary: { [key: string]: string } = {
-	add: "add",
-	activate: "activate",
-	assert	: "assert",
-	remove: "remove",
-	create: "create",
-	count: "count",
-	compare: "compare",
-	collect: "collect",
-	delete: "delete",
-	debounce: "debounce",
-	debounced: "debounced",
-	deactivate: "deactivate",
-	double: "double",
-	expect: "expect",
-	evaluate: "evaluate",
-	get: "get",
-	set: "set",
-	update: "update",
-	reset: "reset",
-	fetch: "fetch",
-	parse: "parse",
-	format: "format",
-	calculate: "calculate",
-	contains: "contains",
-	compute: "compute",
-	validate: "validate",
-	serialize: "serialize",
-	deserialize: "deserialize",
-	filter: "filter",
-	sort: "sort",
-	find: "find",
-	replace: "replace",
-	insert: "insert",
-	save: "save",
-	load: "load",
-	send: "send",
-	receive: "receive",
-	process: "process",
-	handle: "handle",
-	convert: "convert",
-	execute: "execute",
-	compile: "compile",
-	build: "build",
-	render: "render",
-	draw: "draw",
-	write: "write",
-	read: "read",
-	open: "open",
-	close: "close",
-	begin: "begin",
-	end: "end",
-	start: "start",
-	stop: "stop",
-	pause: "pause",
-	resume: "resume",
-	toggle: "toggle",
-	enable: "enable",
-	disable: "disable",
-	configure: "configure",
-	check: "check",
-	invoke: "invoke",
-	launch: "launch",
-	run: "run",
-	apply: "apply",
-	clear: "clear",
-	commit: "commit",
-	rollback: "rollback",
-	encrypt: "encrypt",
-	decrypt: "decrypt",
-	scan: "scan",
-	copy: "copy",
-	move: "move",
-	bind: "bind",
-	unbind: "unbind",
-	grow: "grow",
-	greet: "greet",
-	shrink: "shrink",
-	zoom: "zoom",
-	scale: "scale",
-	flush: "flush",
-	pop: "pop",
-	push: "push",
-	expand: "expand",
-	collapse: "collapse",
-	show: "show",
-	hide: "hide",
-	attach: "attach",
-	detach: "detach",
-	clone: "clone",
-	test: "test",
-	queue: "queue",
-	dequeue: "dequeue",
-	lock: "lock",
-	unlock: "unlock",
-	upload: "upload",
-	download: "download",
-	refresh: "refresh",
-	merge: "merge",
-	split: "split",
-	join: "join",
-	slice: "slice",
-	index: "index",
-	mount: "mount",
-	unmount: "unmount",
-	display: "display",
-	initialize: "initialize",
-	scroll: "scroll",
-	ensure: "ensure",
-	register: "register",
-	say: "say",
-	approve: "approve",
-	decline: "decline",
-	accept: "accept",
-	reject: "reject",
-	request: "request",
-	authorize: "authorize",
-	release: "release",
-	publish: "publish",
-	consume: "consume",
-	submit: "submit",
-	complete: "complete",
-	abort: "abort",
-	confirm: "confirm",
-	verify: "verify",
-	acknowledge: "acknowledge",
-	notify: "notify",
-	normalize: "normalize",
-	listen: "listen",
-	wait: "wait",
-	timeout: "timeout",
-	access: "access",
-	analyze: "analyze",
-	assign: "assign",
-	authenticate: "authenticate",
-	backup: "backup",
-	balance: "balance",
-	bookmark: "bookmark",
-	browse: "browse",
-	broadcast: "broadcast",
-	chat: "chat",
-	cancel: "cancel",
-	capture: "capture",
-	chain: "chain",
-	checkpoint: "checkpoint",
-	compress: "compress",
-	connect: "connect",
-	construct: "construct",
-	coordinate: "coordinate",
-	crawl: "crawl",
-	defragment: "defragment",
-	deny: "deny",
-	deploy: "deploy",
-	derive: "derive",
-	detect: "detect",
-	decrement: "decrement",
-	distribute: "distribute",
-	duplicate: "duplicate",
-	erase: "erase",
-	estimate: "estimate",
-	examine: "examine",
-	exclude: "exclude",
-	extract: "extract",
-	feed: "feed",
-	fix: "fix",
-	flatten: "flatten",
-	forward: "forward",
-	generate: "generate",
-	grant: "grant",
-	has: "has",
-	hash: "hash",
-	identify: "identify",
-	ignore: "ignore",
-	import: "import",
-	include: "include",
-	inspect: "inspect",
-	increment: "increment",
-	integrate: "integrate",
-	intercept: "intercept",
-	interpret: "interpret",
-	interrupt: "interrupt",
-	invalidate: "invalidate",
-	invite: "invite",
-	is: "is",
-	jump: "jump",
-	keep: "keep",
-	learn: "learn",
-	link: "link",
-	log: "log",
-	manage: "manage",
-	map: "map",
-	monitor: "monitor",
-	navigate: "navigate",
-	observe: "observe",
-	optimize: "optimize",
-	organize: "organize",
-	package: "package",
-	perform: "perform",
-	preserve: "preserve",
-	prevent: "prevent",
-	provide: "provide",
-	profile: "profile",
-	purge: "purge",
-	query: "query",
-	quickstart: "quickstart",
-	relay: "relay",
-	rename: "rename",
-	report: "report",
-	resolve: "resolve",
-	retry: "retry",
-	revert: "revert",
-	rotate: "rotate",
-	schedule: "schedule",
-	score: "score",
-	search: "search",
-	select: "select",
-	settle: "settle",
-	share: "share",
-	shield: "shield",
-	sign: "sign",
-	synchronize: "synchronize",
-	terminate: "terminate",
-	translate: "translate",
-	transport: "transport",
-	track: "track",
-	trim: "trim",
-	undo: "undo",
-	upgrade: "upgrade",
-	visit: "visit",
-	watch: "watch",
-	welcome: "welcome",
-	wipe: "wipe",
-	zip: "zip",
-};
