--- server/src/services/analysis/index.ts
@@ -1,1 +1,0 @@
-export * from "./commentAnalyzer";
