--- server/src/services/validation/index.ts
@@ -1,2 +1,0 @@
-export * from "./nameValidator";
-export * from "./themeValidator";
